Moving remaining device communication in hardwareWallet package
