+++ src/api/cli/device_check_message_signature.go
@@ -1,12 +1,7 @@
 package cli
 
 import (
-	"fmt"
-
-	"github.com/golang/protobuf/proto"
-	messages "github.com/skycoin/skycoin/protob"
 	hardwareWallet "github.com/skycoin/skycoin/src/hardware-wallet"
-	"github.com/wire"
 	gcli "github.com/urfave/cli"
 )
 
@@ -35,26 +30,7 @@ func deviceCheckMessageSignatureCmd() gcli.Command {
 			message := c.String("message")
 			signature := c.String("signature")
 			address := c.String("address")
-			dev, _ := hardwareWallet.GetTrezorDevice()
-
-			skycoinCheckMessageSignature := &messages.SkycoinCheckMessageSignature{
-				Address:   proto.String(address),
-				Message:   proto.String(message),
-				Signature: proto.String(signature),
-			}
-
-			data, _ := proto.Marshal(skycoinCheckMessageSignature)
-
-			chunks := hardwareWallet.MakeTrezorMessage(data, messages.MessageType_MessageType_SkycoinCheckMessageSignature)
-
-			for _, element := range chunks {
-				_, _ = dev.Write(element[:])
-			}
-
-			var msg wire.Message
-			msg.ReadFrom(dev)
-
-			fmt.Printf("Success %d! address that issued the signature is: %s\n", msg.Kind, msg.Data)
+			hardwareWallet.DeviceCheckMessageSignature(message, signature, address)
 		},
 	}
 }
